Refactore structure of security package
