+++ gaminatorium-main/monolith/src/main/java/eu/gaminatorium/security/credentials/GaminatoriumAuthorities.java
@@ -1,4 +1,4 @@
-package eu.gaminatorium.security;
+package eu.gaminatorium.security.credentials;
 
 import lombok.AllArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
