+++ gaminatorium-main/monolith/src/main/java/eu/gaminatorium/security/credentials/UserSecurity.java
@@ -1,4 +1,4 @@
-package eu.gaminatorium.security;
+package eu.gaminatorium.security.credentials;
 
 import eu.gaminatorium.user.User;
 import lombok.AllArgsConstructor;
